# apps/web/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MoonPlants API",
    "version": "1.0.0",
    "description": "REST API for MoonPlants — smart plant monitoring and automated watering system. Consists of two namespaces:\n\n- **`/api/v1/`** — User-facing endpoints (Supabase JWT auth)\n- **`/api/iot/v1/`** — Device-facing endpoints (HMAC-SHA256 auth)",
    "contact": {
      "name": "MoonPlants"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    },
    {
      "url": "https://your-production-domain.com",
      "description": "Production (replace with actual domain)"
    }
  ],
...
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "example": "Unauthorized"
          }
        }
      },
...
      "Plant": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "owner_user_id": {
            "type": "string",
            "format": "uuid"
          },
          "device_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "example": "Monstera Deliciosa"
          },
          "species_name": {
            "type": "string"
          },
          "species_cache_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "soil_channel": {
            "type": [
              "integer",
              "null"
            ]
          },
          "image_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "image_source": {
            "type": [
              "string",
              "null"
            ],
            "enum": [
              "user",
              "perenual",
              "none"
            ]
          },
          "auto_watering_enabled": {
            "type": "boolean"
          },
          "auto_light_enabled": {
            "type": "boolean"
          },
          "age_months": {
            "type": [
              "integer",
              "null"
            ]
          },
          "pot_height_cm": {
            "type": [
              "number",
              "null"
            ]
          },
          "pot_diameter_cm": {
            "type": [
              "number",
              "null"
            ]
          },
          "pot_volume_ml": {
            "type": [
              "integer",
              "null"
            ]
          },
          "last_watered_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreatePlantRequest": {
        "type": "object",
        "required": [
          "name",
          "speciesName"
        ],
        "properties": {
          "deviceId": {
            "type": "string",
            "format": "uuid",
            "description": "Optional: link to a claimed device"
          },
          "soilChannel": {
            "type": "integer",
            "minimum": 1,
            "maximum": 16,
            "description": "Device channel index (required when deviceId set)"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200,
            "example": "My Monstera"
          },
          "speciesName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 300,
            "example": "Monstera deliciosa"
          },
          "perenualId": {
            "type": "integer",
            "description": "Perenual species ID to auto-fetch image/data"
          },
          "potVolumeMl": {
            "type": "integer",
            "minimum": 1
          },
          "potDiameterCm": {
            "type": "number",
            "minimum": 0
          },
          "potHeightCm": {
            "type": "number",
            "minimum": 0
          },
          "lastWateredAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        }
      },
...
      "Prediction": {
        "type": "object",
        "properties": {
          "plant_id": {
            "type": "string",
            "format": "uuid"
          },
          "predicted_at": {
            "type": "string",
            "format": "date-time"
          },
          "next_watering_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "confidence": {
            "type": [
              "number",
              "null"
            ],
            "minimum": 0,
            "maximum": 1
          },
          "reason": {
            "type": [
              "string",
              "null"
            ]
          },
          "should_water_now": {
            "type": "boolean"
          }
        }
      },
...
      "ClaimDeviceRequest": {
        "type": "object",
        "required": [
          "deviceId",
          "claimCode"
        ],
        "properties": {
          "deviceId": {
            "type": "string",
            "format": "uuid"
          },
          "claimCode": {
            "type": "string",
            "minLength": 16,
            "maxLength": 64,
            "example": "CLAIM-XXXX-YYYY-ZZZZ"
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supabase JWT access token. Pass via Authorization: Bearer <token>"
      },
      "IotHmac": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Device-Signature",
        "description": "HMAC-SHA256 device signature. IoT endpoints require all 5 headers: X-Device-Id, X-Device-Seq, X-Device-Timestamp, X-Content-SHA256, X-Device-Signature. Canonical string = METHOD\\nPATH\\nX-Device-Id\\nX-Device-Seq\\nX-Device-Timestamp\\nX-Content-SHA256"
      }
    }
  },
  "paths": {
...
    "/api/v1/devices/claim": {
      "post": {
        "tags": [
          "Devices"
        ],
        "summary": "Claim a device",
        "description": "Links an unclaimed device to the authenticated user's account using the device's claim code.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClaimDeviceRequest"
              },
              "examples": {
                "claim": {
                  "summary": "Claim with valid code",
                  "value": {
                    "deviceId": "e2a3b1c4-0000-0000-0000-000000000001",
                    "claimCode": "CLAIM-ABCD-EFGH-IJKL"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Device claimed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "deviceId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Device not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Device already claimed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many failed attempts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/plants": {
      "post": {
        "tags": [
          "Plants"
        ],
        "summary": "Create a plant",
        "description": "Creates a new plant and optionally links it to a device channel. If a Perenual ID is supplied, species data and image are auto-fetched.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlantRequest"
              },
              "examples": {
                "withDevice": {
                  "summary": "Plant linked to device channel 1",
                  "value": {
                    "deviceId": "e2a3b1c4-0000-0000-0000-000000000001",
                    "soilChannel": 1,
                    "name": "My Monstera",
                    "speciesName": "Monstera deliciosa",
                    "perenualId": 1882,
                    "potVolumeMl": 3000
                  }
                },
                "standalone": {
                  "summary": "Standalone plant without device",
                  "value": {
                    "name": "Cactus",
                    "speciesName": "Echinocactus grusonii"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Plant created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "plant": {
                      "$ref": "#/components/schemas/Plant"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or channel out of range",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Device not owned by user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Channel already in use",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/plants/{plantId}/predictions": {
      "get": {
        "tags": [
          "Plants"
        ],
        "summary": "Get watering prediction",
        "description": "Runs the rule-based prediction model and returns the next watering recommendation.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "plantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Prediction result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "prediction": {
                      "$ref": "#/components/schemas/Prediction"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Plant not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}